+++ src/consumer/consumer.js
@@ -721,7 +721,7 @@ export default class Consumer {
     if (!track && removedIds.bitIds) {
       const bitMap = await this.getBitMap();
       removedIds.bitIds.forEach((id) => {
-        bitMap.removeComponent(BitId.parse(id));
+        bitMap.removeComponent(id);
       });
       await bitMap.write();
     }