+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- fixed remove of exported component 
+
 ## [0.10.9-dev.4] - 2017-10-18
 
 - typescript - extract only relevant dependencies from link files (files that only require other files)
@@ -15,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - prevent bare-scope corruption when the export process fails
 - fixed stderr maxBuffer exceeded bug in ci-update cmd
 - fixed - writing dist files to wrong directory during bit tag / test commands
+
 
 ## [0.10.9-dev.3] - 2017-10-10